+++ src/NServiceBus.Core/EndpointNotifications.cs
@@ -2,7 +2,7 @@ namespace NServiceBus
 {
     using System;
     using System.Collections.Generic;
-    using NServiceBus.Routing;
+    using NServiceBus.EndpointControl;
 
     /// <summary>
     ///     Endpoint notifications
@@ -13,14 +13,14 @@ namespace NServiceBus
         ///     Notification when an endpoint has been idled for a configured amount of time
         ///     and it should be safe to shut it down without leaving a backlog of messages.
         /// </summary>
-        public IObservable<SafeDisconnect> SafeToDisconnect
+        public IObservable<EndpointSafeToDisconnected> SafeToDisconnect
         {
             get { return disconnectEventList; }
         }
 
         internal void InvokeSafeToDisconnect(Dictionary<string, string> headers)
         {
-            disconnectEventList.Publish(new SafeDisconnect(new Dictionary<string, string>(headers)));
+            disconnectEventList.Publish(new EndpointSafeToDisconnected(new Dictionary<string, string>(headers)));
         }
 
         void IDisposable.Dispose()
@@ -28,6 +28,6 @@ namespace NServiceBus
             // Injected
         }
 
-        Observable<SafeDisconnect> disconnectEventList = new Observable<SafeDisconnect>();
+        Observable<EndpointSafeToDisconnected> disconnectEventList = new Observable<EndpointSafeToDisconnected>();
     }
 }
